Use time.Duration in TimerStats.

+++ termite/fsserver.go
@@ -28,8 +28,8 @@ func (me *FsServer) FileContent(req *cba.ContentRequest, rep *cba.ContentRespons
 	start := time.Now()
 	err := me.contentCache.Serve(req, rep)
 	dt := time.Now().Sub(start)
-	me.stats.Log("FsServer.FileContent", int64(dt))
-	me.stats.LogN("FsServer.FileContentBytes", int64(len(rep.Chunk)), int64(dt))
+	me.stats.Log("FsServer.FileContent", dt)
+	me.stats.LogN("FsServer.FileContentBytes", int64(len(rep.Chunk)), dt)
 	return err
 }
 
@@ -49,6 +49,6 @@ func (me *FsServer) GetAttr(req *AttrRequest, rep *AttrResponse) error {
 	}
 	rep.Attrs = append(rep.Attrs, a)
 	dt := time.Now().Sub(start)
-	me.stats.Log("FsServer.GetAttr", int64(dt))
+	me.stats.Log("FsServer.GetAttr", dt)
 	return nil
 }
