+++ termite/mirror.go
@@ -249,8 +249,8 @@ func (me *Mirror) FileContent(req *cba.ContentRequest, rep *cba.ContentResponse)
 	start := time.Now()
 	err := me.worker.contentCache.Serve(req, rep)
 	dt := time.Now().Sub(start)
-	me.rpcFs.timings.Log("Mirror.FileContent", int64(dt))
-	me.rpcFs.timings.LogN("Mirror.FileContentBytes", int64(len(rep.Chunk)), int64(dt))
+	me.rpcFs.timings.Log("Mirror.FileContent", dt)
+	me.rpcFs.timings.LogN("Mirror.FileContentBytes", int64(len(rep.Chunk)), dt)
 
 	return err
 }
